requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.41
identify==2.6.1
Markdown==3.7
mimesis==18.0.0
nodeenv==1.9.1
phonenumberslite==8.13.46
platformdirs==4.3.6
pre-commit==3.8.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==69.0.3
smmap==5.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.5
